# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reading Quest API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Profile {
  id: number;
  name: string;
  gems: number;
  stars: number;
  petLevel: number;
  petXp: number;
  petXpForNextLevel: number;
  petXpProgressPercent: number;
}

export interface World {
  id: number;
  slug: string;
  name: string;
  blurb: string;
  /** 1=gentle, 2=breezy, 3=brave */
  difficulty: number;
  difficultyLabel: string;
  thumbnailUrl: string;
  accentColor: string;
  chipColor: string;
  chipTextColor: string;
}

export interface Story {
  id: number;
  worldId: number;
  slug: string;
  title: string;
  summary: string;
  chapterCount: number;
  finishedChapterCount: number;
}

export interface ChapterSummary {
  id: number;
  storyId: number;
  sortIndex: number;
  title: string;
  finished: boolean;
}

export interface StoryDetail {
  id: number;
  worldId: number;
  title: string;
  summary: string;
  chapters: ChapterSummary[];
}

export interface TappableWord {
  key: string;
  word: string;
  syllables: string[];
  reaction: string;
}

export interface Chapter {
  id: number;
  storyId: number;
  sortIndex: number;
  title: string;
  sceneImageUrl: string;
  chapterCount: number;
  /** @nullable */
  nextChapterId?: number | null;
  paragraphs: string[];
  tappableWords: TappableWord[];
}

export interface StartSessionRequest {
  chapterId: number;
}

export interface Session {
  id: number;
  profileId: number;
  chapterId: number;
  storyId: number;
  startedAt: string;
  activeMs: number;
  /** active|paused|finished */
  status: string;
}

export interface HeartbeatRequest {
  /** Milliseconds of active reading since the last heartbeat (clamped server-side) */
  activeMsDelta: number;
}

export interface SessionRewards {
  session: Session;
  gemsAwarded: number;
  starsAwarded: number;
  xpAwarded: number;
  petLeveledUp: boolean;
  /** @nullable */
  newPetLevel?: number | null;
  profile: Profile;
}

export interface WordHelpRequest {
  wordKey: string;
}

export interface WordHelp {
  word: TappableWord;
  /** Number of times this kid has asked for help on this word */
  helpCount: number;
}

export interface PetState {
  profileId: number;
  level: number;
  xp: number;
  xpForNextLevel: number;
  xpProgressPercent: number;
  /** 0-100 */
  fullness: number;
  /** 0-100 */
  happiness: number;
  /** ecstatic|happy|okay|lonely */
  mood: string;
  /** @nullable */
  equippedHat?: string | null;
  glowColor: string;
  decor: string[];
  gems: number;
  stars: number;
}

export interface PetItemRequest {
  itemId: string;
}

export interface EquipRequest {
  /** hat|glow */
  slot: string;
  /** @nullable */
  itemId?: string | null;
}

export interface DecorToggleRequest {
  itemId: string;
  on: boolean;
}

export interface ShopItem {
  id: string;
  /** snack|hat|glow|decor */
  kind: string;
  name: string;
  description: string;
  gemPrice: number;
  fullnessBoost: number;
  happinessBoost: number;
  requiresPetLevel: number;
  owned: boolean;
  locked: boolean;
  /** @nullable */
  lockHint?: string | null;
  emoji: string;
  /** @nullable */
  glowColor?: string | null;
}

export interface PurchaseRequest {
  itemId: string;
}

export interface GrownupsAuthRequest {
  passcode: string;
}

export interface GrownupsAuthResult {
  token: string;
}

export interface GrownupsSummary {
  minutesRead: number;
  storiesFinished: number;
  chaptersFinished: number;
  wordsHelped: number;
  sessionsCount: number;
}

export interface DailyMinutes {
  /** YYYY-MM-DD */
  date: string;
  minutes: number;
}

export interface WordHelpAggregate {
  wordKey: string;
  word: string;
  helpCount: number;
  lastSeenAt: string;
}

export interface FinishedStory {
  storyId: number;
  title: string;
  worldName: string;
  finishedAt: string;
}

export interface ActivityEvent {
  id: number;
  /** session_started|chapter_finished|word_help|pet_fed|item_purchased|item_equipped */
  kind: string;
  label: string;
  createdAt: string;
}

/**
 * @nullable
 */
export type GetActiveSession200 = {
  id: number;
  profileId: number;
  chapterId: number;
  storyId: number;
  startedAt: string;
  activeMs: number;
  status: string;
} | null;