# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Reading Quest API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pet/feed:
    post:
      operationId: feedPet
      tags: [pet]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PetItemRequest" }
      responses:
        "200":
          description: Updated pet state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PetState" }
        "400":
          description: Bad request (e.g. not enough gems)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /pet/equip:
    post:
      operationId: equipPetItem
      tags: [pet]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EquipRequest" }
      responses:
        "200":
          description: Updated pet state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PetState" }
        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /pet/decor:
    post:
      operationId: togglePetDecor
      tags: [pet]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DecorToggleRequest" }
      responses:
        "200":
          description: Updated pet state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PetState" }
        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
  /shop/purchase:
    post:
      operationId: purchaseShopItem
      tags: [pet]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PurchaseRequest" }
      responses:
        "200":
          description: Updated pet state after purchase
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PetState" }
        "400":
          description: Bad request (e.g. not enough gems, locked, already owned)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /grownups/auth:
    post:
      operationId: grownupsAuth
      tags: [grownups]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GrownupsAuthRequest" }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GrownupsAuthResult" }
        "401":
          description: Wrong passcode
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /grownups/summary:
    get:
      operationId: grownupsSummary
      tags: [grownups]
      parameters:
        - in: header
          name: x-grownup-token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Top-line totals for the past 30 days
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GrownupsSummary" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /grownups/weekly-minutes:
    get:
      operationId: grownupsWeeklyMinutes
      tags: [grownups]
      parameters:
        - in: header
          name: x-grownup-token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Reading minutes per day for the last 7 days
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DailyMinutes" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /grownups/words:
    get:
      operationId: grownupsWords
      tags: [grownups]
      parameters:
        - in: header
          name: x-grownup-token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Word help aggregation
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WordHelpAggregate" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /grownups/finished-stories:
    get:
      operationId: grownupsFinishedStories
      tags: [grownups]
      parameters:
        - in: header
          name: x-grownup-token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Finished stories with timestamps
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FinishedStory" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /grownups/recent-activity:
    get:
      operationId: grownupsRecentActivity
      tags: [grownups]
      parameters:
        - in: header
          name: x-grownup-token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Most recent 25 events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ActivityEvent" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    PetState:
      type: object
      properties:
        profileId: { type: integer }
        level: { type: integer }
        xp: { type: integer }
        xpForNextLevel: { type: integer }
        xpProgressPercent: { type: integer }
        fullness: { type: integer, description: "0-100" }
        happiness: { type: integer, description: "0-100" }
        mood: { type: string, description: "ecstatic|happy|okay|lonely" }
        equippedHat:
          type: ["string", "null"]
        glowColor: { type: string }
        decor:
          type: array
          items: { type: string }
        gems: { type: integer }
        stars: { type: integer }
      required: [profileId, level, xp, xpForNextLevel, xpProgressPercent, fullness, happiness, mood, glowColor, decor, gems, stars]
    PetItemRequest:
      type: object
      properties:
        itemId: { type: string }
      required: [itemId]
    EquipRequest:
      type: object
      properties:
        slot: { type: string, description: "hat|glow" }
        itemId:
          type: ["string", "null"]
      required: [slot]
    DecorToggleRequest:
      type: object
      properties:
        itemId: { type: string }
        on: { type: boolean }
      required: [itemId, on]
...
    PurchaseRequest:
      type: object
      properties:
        itemId: { type: string }
      required: [itemId]
    GrownupsAuthRequest:
      type: object
      properties:
        passcode: { type: string }
      required: [passcode]
    GrownupsAuthResult:
      type: object
      properties:
        token: { type: string }
      required: [token]
    GrownupsSummary:
      type: object
      properties:
        minutesRead: { type: integer }
        storiesFinished: { type: integer }
        chaptersFinished: { type: integer }
        wordsHelped: { type: integer }
        sessionsCount: { type: integer }
      required: [minutesRead, storiesFinished, chaptersFinished, wordsHelped, sessionsCount]
    DailyMinutes:
      type: object
      properties:
        date: { type: string, description: "YYYY-MM-DD" }
        minutes: { type: integer }
      required: [date, minutes]
    WordHelpAggregate:
      type: object
      properties:
        wordKey: { type: string }
        word: { type: string }
        helpCount: { type: integer }
        lastSeenAt: { type: string, format: date-time }
      required: [wordKey, word, helpCount, lastSeenAt]
    FinishedStory:
      type: object
      properties:
        storyId: { type: integer }
        title: { type: string }
        worldName: { type: string }
        finishedAt: { type: string, format: date-time }
      required: [storyId, title, worldName, finishedAt]
    ActivityEvent:
      type: object
      properties:
        id: { type: integer }
        kind: { type: string, description: "session_started|chapter_finished|word_help|pet_fed|item_purchased|item_equipped" }
        label: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, kind, label, createdAt]