# lib/api-zod/src/generated/types/finishedStory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reading Quest API
 * OpenAPI spec version: 0.1.0
 */

export interface FinishedStory {
  storyId: number;
  title: string;
  worldName: string;
  finishedAt: Date;
}